폴더 경로 수정
mypage 폴더 경로 수정
