--- src/main/java/org/flower/controllers/user/mypage/MyPageController.java
@@ -1,5 +1,0 @@
-package org.flower.controllers.user.mypage;
-
-
-public class MyPageController {
-}
