+++ src/main/java/org/flower/controllers/user/mypage/profile/ProfileController.java
@@ -1,14 +1,15 @@
-package org.flower.controllers.mypage;
+package org.flower.controllers.user.mypage.profile;
+
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
-@RequestMapping("/mypage")
-public class MypageHomeController {
+@RequestMapping("/mypage/profile/profilePage")
+public class ProfileController {
     @GetMapping
     public String main(){
-        return "/mypage/main/home";
+        return "/mypage/profile/profilePage";
     }
 }
