+++ src/main/resources/templates/front/mypage/main/home.html
@@ -8,7 +8,7 @@
 <br>
 <main layout:fragment="content">
     <h1>마이페이지 메인페이지입니다.</h1>    
-    <a th:href="@{/mypage/profile/profilePage}">
+    <a th:href="@{/templates/front/mypage/profile/profilePage}">
         <button>프로필 사진 보기</button>
     </a>
 </main>
