+++ src/main/resources/templates/front/mypage/outlines/footer.html
